requirements.txt
fastapi[standard]==0.115.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
